+++ spec/components/schemas/content/bases/buttons/calendar-event.ts
@@ -1,5 +1,5 @@
 import { SchemaObject } from 'openapi3-ts';
-import { createComponentRef } from '../../../../../utils/ref';
+import { createComponentRef } from '../../../../../../utils/ref';
 import { ref as baseRef } from './base';
 
 const calendarevent: SchemaObject = {
