+++ spec/components/schemas/content/bases/card-media-base.ts
