+++ spec/components/schemas/content/bases/quick-reply-list.ts
@@ -1,6 +1,6 @@
 import { SchemaObject } from 'openapi3-ts';
 import { createComponentRef } from '../../../../../utils/ref';
-import { ref as buttonsRef } from './all-buttons';
+import { ref as buttonsRef } from './buttons/all';
 
 export const quickReplyList: SchemaObject = {
   type: 'array',
